Refactors boilerplate
