+++ contentcuration/contentcuration/utils/recommendations.py
@@ -1,3 +1,5 @@
+from typing import Union
+
 from automation.utils.appnexus.base import Adapter
 from automation.utils.appnexus.base import Backend
 from automation.utils.appnexus.base import BackendFactory
@@ -46,3 +48,22 @@ class RecommendationsAdapter(Adapter):
     def get_recommendations(self) -> RecommendationsResponse:
         request = RecommedationsRequest()
         return self.backend.make_request(request)
+
+
+class Recommendations(Backend):
+
+    def connect(self) -> None:
+        return super().connect()
+
+    def make_request(self, request) -> Union[EmbeddingsResponse, RecommendationsResponse]:
+        return super().make_request(request)
+
+    def request(self) -> None:
+        return super().request()
+
+    def response(self) -> None:
+        return super().response()
+
+    @classmethod
+    def _create_instance(cls) -> 'Recommendations':
+        return cls()
